Add data reader file
